several updates

+++ spark-setup/0_startdns.sh
@@ -1,5 +1,13 @@
-# TODO : include check if container exist before issuing stop and rm commands
-docker stop dns
-docker rm dns
+containerExists=`docker ps -a | grep "dns"`
+
+if [ "$containerExists" ];then
+
+	docker stop dns
+	#Remove the container
+	docker rm dns
+fi
+
+
+
 docker run -d -t --name dns -h dns.example.com -v /etc/bind pti1/bind9:secondversion
